Refactor: Move all setExpenditureState helpers into separate file
