+++ src/redux/sagas/utils/index.ts
@@ -11,9 +11,8 @@ export * from './metadataChangelog.ts';
 export * from './ipfs.ts';
 export * from './annotations.ts';
 export * from './proofs.ts';
-export * from './expenditureMulticallData.ts';
 export * from './expenditures.ts';
-export * from './expendituresMulticall.ts';
+export * from './setExpenditureStateHelpers.ts';
 
 export { updateDomainReputation } from './updateDomainReputation.ts';
 export { createActionMetadataInDB } from './createActionMetadata.ts';
